feat: refactor post page and remove unused components
Updated the post page to navigate to the root instead of the post path. Removed the unused Index component and the post page file, streamlining the application structure and improving code maintainability.
